discord-bot/requirements.txt
discord.py>=2.5.2
python-dotenv
pytz
requests
Pillow==10.0.0
aiohttp==3.8.5
asyncpg==0.28.0
beautifulsoup4==4.12.2
psutil==5.9.5
PyNaCl==1.5.0
wavelink==2.6.4
humanize==4.7.0
colorlog==6.7.0
motor==3.2.0
dnspython==2.4.2
python-dateutil==2.8.2
supabase>=2.0.0
httpx>=0.24.0 